+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 * jdbc: let tryLock/unlock not eat exceptions; also log if unlock returns false
+* jobs: fix locking/unlocking of non-transactional jobs that should not run in parallel
 
 # 2.0.3
 * server: AppScope.async will now log any exceptions to avoid silent failures